% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_report.R
\name{genomic_report}
\alias{genomic_report}
\title{Genomics Lab Report}
\usage{
genomic_report(
  title = "Genetics Lab Report",
  author = Sys.info()[["user"]],
  date = format(Sys.time(), "\%B \%d, \%Y")
)
}
\arguments{
\item{title}{a string, the title of the report. Defaults to "Genetics Lab Report"}

\item{author}{a string, the name of the author. Defaults to system username}

\item{date}{a string. Defaults to the system date}
}
\value{
file name of the outputted report, which is a word-document
}
\description{
genomic_report() creates a a full laboratory report
}
\examples{
genomic_report("My Report", "John Doe")
}
