% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_summary.R
\name{genomic_summary}
\alias{genomic_summary}
\title{Genomic Summary}
\usage{
genomic_summary(data, log2, sig)
}
\arguments{
\item{data}{name of data set}

\item{log2}{name of column with log2 values}

\item{sig}{column name in data set that states changes were significant}
}
\description{
This function provides some basic information on the data set. It first
provides a table of total number of significantly deferentially expressed
genes. Then in runs a binomial exact test on those numbers. Lastly, it
returns a basic histogram of log2 (fold change) values in order for the
user to gain an understanding of the skew.
}
\examples{
genomic_summary(data = gendata, log2 = "log2.fold_change.", sig = "significant")
}
