% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_tables.R
\name{genomic_tables}
\alias{genomic_tables}
\title{Genomic Tables}
\usage{
genomic_tables(
  data,
  U.D.diffexp = c("neither", "up", "down"),
  gene.n,
  log2,
  pval,
  sig,
  num.rows = 10
)
}
\arguments{
\item{data}{name of data set}

\item{U.D.diffexp}{pick either "up" or "down" to filter for genes in those categories}

\item{gene.n}{column of data set with gene names}

\item{log2}{column of data set with log2 values}

\item{pval}{column of data set with p-values}

\item{sig}{column name in data set that states changes were significant}

\item{num.rows}{number of rows to display in the table}
}
\value{
returns a huxtable
}
\description{
Generates tables of gene names, log2 values, p-values. One can filter for genes
that were up or down regulated, or the default is neither. Function will prompt
for the number of rows you wish to view.
}
\examples{
genomic_tables(data = gendata, U.D.diffexp = "up", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")
genomic_tables(data = gendata, U.D.diffexp = "down", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")
genomic_tables(data = gendata, U.D.diffexp = "neither", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")
genomic_tables(data = gendata, gene.n = "gene", log2 = "log2.fold_change.",
        pval = "p_value", sig = "significant")
}
