% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import dataset}
\usage{
import_data(file, ...)
}
\arguments{
\item{file}{Path to data file.}

\item{...}{Additional arguments.}
}
\value{
A dataframe containing the dataset
}
\description{
import_data() lets the user import a dataset from given file path.
}
\details{
This function lets the user import dataset into R from Stata, SPSS, SAS,
excel, and delimited text files and returns a dataframe.
}
\examples{
\dontrun{import("./filename.csv")}

}
